+++ tests/testthat/test_fanout.R
@@ -2,10 +2,10 @@ context("Fanout execution")
 
 test_that("fan out IDs are correct", {
 
-  sims <- c(5, 1, 2, 1, 1, 2)
+  sims <- c(5, 2, 1, 2)
   design <- two_arm_designer(rho = 0)
   
   sx <- simulate_design(design, sims = sims)
   
-  expect_equivalent(apply(sx[,c("step_1_draw","step_3_draw","step_6_draw")], 2, max), c(5, 10, 20))
+  expect_equivalent(apply(sx[,c("step_1_draw","step_2_draw","step_4_draw")], 2, max), c(5, 10, 20))
 })